+++ server/doc.go
@@ -1,0 +1,2 @@
+// Package server предоставляет фасад для быстрого запуска HTTP-сервисов на базе nope.
+package server